backend/requirements.txt
click==8.1.3
Flask==2.2.0
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
zipp==3.8.1
